+++ CHANGELOG.rst
@@ -1,7 +1,7 @@
 Change Log
 ----------
 
-0.1.dev2 (2015-01-11)
+0.1.dev2 (2015-01-14)
 +++++++++++++++++++++
 
 **Features**
@@ -9,6 +9,7 @@ Change Log
 - Added ``altapay.Transaction`` and the ability to find a transaction by its transaction ID in the AltaPay service
 - Added ``altapay.Transaction.capture()`` which captures a transaction that has already been loaded. Optinally, parameters can be passed which allows for partial captures (see the AltaPay documentation for full list of possible arguments)
 - Added a public facing API for converting an AltaPay XML response (as a string) to a Python dictionary (``altapay.utils.xml_to_dict``)
+- Added ``altapay.Callback`` which wraps a callback response from AltaPay, and automatically wraps the coupled transactions in ``altapay.Transaction`` objects
 
 **Bugfixes**